requirements.txt
Flask
Flask-RESTX
Flask-SQLAlchemy
psycopg2-binary
alembic
paho-mqtt
gunicorn